serialize and deserialize with passport, add bcryptjs
