+++ db/index.js
@@ -1,6 +1,34 @@
-const db = require('./models')
-    , User = require('./models/user')
+const app = require('../')
+    , name = app.env.DATABASE_NAME || app.name
+    , url = app.env.DATABASE_URL || `postgres://localhost:5432/${name}`
+    , Sequelize = require('Sequelize')
+    , sync = app.env.NODE_DEV === 'development' ? { force: true } : {}
 
-// Define association
+const db = module.exports = new Sequelize(url, {
+  define: {
+    freezeTableName: true,
+    timestamps: true,
+    underscored: true
+  },
+  logging: false
+})
 
-module.exports = db
+db.syncAndLaunch = (callback) => {
+  const attemptSync = (reattempt = false) => {
+    db.sync(sync)
+    .then(callback)
+    .catch((err) => {
+      console.error('failed to sync database')
+      if (!reattempt) {
+        console.log('reattempting start up')
+        return new Promise((res, _) => {
+          require('child_process').exec(`createdb "${name}"`, res)
+        })
+        .then(() => attemptSync(true))
+      } else {
+        console.error('failed to reattempt sync', err)
+      }
+    })
+  }
+  return attemptSync
+}
